+++ packages/adminui-vue/CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - `UBaseInput`,`UAutoField`,`USelectEntity`: fixed missing injection in case when the control is outside the
+   `UFormContainer`
 
 ## [1.13.21] - 2020-11-20
 ### Fixed